[css value tracing] Resolve percentages in shorthands
For this to work we need to keep track of the position of a top-level
AST node directly under the declaration's fully-substituted property
value. For tracing, we need to thread this position through the
hierarchy of substitutions as well so that we can additionally determine
the position of top-level AST nodes directly under the var (either the
fallback value's nodes or the property value nodes from the custom
property declaration we're substituting in).

Fixed: 401213719
Change-Id: I6855dc7a9420caa91979c34e32ad64e4b2ae58ab
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6428570
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>