RPP: add support for `nodeName` in PaintImage events
This CL adds support for `nodeName` to Paint & PaintImage events, and
uses the Node name in the information given to the LLM for the CLS
Insights AI integration.

There is much more we can do with this property; we can use it when we
don't have a fresh trace to output this in cases where we can't ask CDP
for a proper name, but this is the first step.

Backend support for this landed in crrev.com/c/6491448.

R=andoli@chromium.org

Bug: 413284569
Change-Id: I18b6f73cce0baa23d89c4f8443c7e752b1e8072b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6491415
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>