[ExtensionAPI] Expose 'buildId' on 'Resource'
The buildId/debugId is reported via CDP scriptParsed events and we
stash it on the SDK Script object.

Later on, when we report Resources to extensions, we resolve the
UISourceCode back to the underlying script. Note that this
relationship is 1:n. E.g. worker that use the same script get
de-duplicated into a single UISourceCode. We take the first one
with a buildId on it, which should be fine in 99% of cases.

R=pfaffe@chromium.org

Bug: 413177407
Change-Id: I63641a7f76cd8dd4a985172594e7dc982cebcbcd
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6491310
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>