RPP: disable `v8.compile` category
It adds lots of overhead and a lot of relatively noisy events
(CompileCode) that we don't think provide much value.

So, the plan is to disable it and see if we get feedback about the
events now missing. If we do, we can bring back those events either as
part of this category (perhaps via a setting to enable the category) or
move the specific event(s) that users miss to another category to
minimise the overhead.

R=paulirish@chromium.org

Bug: 414330508
Change-Id: I9f03b01a498d80ca33d1e025ac7e815f59ffdc37
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6494747
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>