Always download emscripten if binary isn't found
Regardless if the `build-revision` content is current

This prevents a bug happenig in the bots where the emscripten binaries aren't present but also aren't downloaded by the gclient hooks because the script bails out if the build-revision is current.

Although the underlying reason behind the binaries not being present when the build-revision file is unclear, this serves as immediate remedy to prevent blocking build failures.

Bug: none
Change-Id: I5df1f229de6e0686917011dea29ca2189d73b216
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6494848
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>