[breakpoint] Don't spam the backend when enabling Debugger takes a while
There are sites where the 'Debugger.enable' CDP call can take a couple
of seconds. During this time, the BreakpointManager currently tries to
repeatedly set the breakpoint even though the call is guaranteed to
fail.

This CL changes to logic to wait for the 'Debugger' domain to enable
after the first attempt fails.

We don't want to generally wait for the enable result as that could
cause us to miss breakpoints. In most cases sending off the
'Debugger.enable' call and already send 'Debugger.setBreakpointByUrl'
while the enable is still in-flight, is fine.

R=kimanh@chromium.org

Bug: 391200159
Change-Id: Ib7f40aa85d5fdcb31d45898d4e070478b79fa890
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6468180
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>