Improve DomFragment `initializer`, `references` and `replacementLocation` semantics.
Made initializer point to the value with which the value is being initialized, not its the identifier at the point of initialization. This is what we need anyway, but because member and local vars are initialized differently, we didn't handle member vars correctly.

Also made `initializer` and `references` to update from ClassMember on read, instead of on `DomFragment.getOrCreate`, which was also causing subtle bugs.

Finally turned `replacementLocation` into a `replacer` function and moved the Widget-specific logic in the corresponding file. Without this we were treating any member variable as widget element.


Bug: 414535492
Change-Id: I4e40fa93af6a6be85b88e771f9a811e6cd35a23d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6494592
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>