Fix "What's new" not scrolling
This was caused by one of the changes in crrev.com/c/6441440 which made
the element a panel. I didn't want to undo that as that change makes
sense.

When looking at the CSS, I noticed a lot of flex elements. Flex +
overflow isn't always a nice combo because of how Flex shrinks + expands
its children. By removing the flexing from the list of "What's new",
this allowed the list to take up the right height, and then we can style
the container to have `overflow: auto` and take up the space of the
panel.

R=dsv@chromium.org

Bug: 414948533
Change-Id: I347917b73310ac6e90261784ba8f4273b5cc3f59
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6504357
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nikolay Vitkov <nvitkov@chromium.org>