[css value tracing] Count longhands in outermost property value
To determine whether we can expand a percentage in a shorthand we need
to check whether the property declaration sets sufficiently many
longhand values to distinguish width and height. For that we need to
count longhands in the outermost property value and no for the
subexpression we run tracing on.

Bug: 396080529
Change-Id: Ibc3aa6018e0e24b16f52042bf2e4fc8c2e6c8a6a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6494578
Reviewed-by: Eric Leese <leese@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Auto-Submit: Philip Pfaffe <pfaffe@chromium.org>