[ui] Make 'widgetConfig' more flexible and cleanup types
This CL changes the first argument to `widgetConfig` to also take
functions that produce 'Widget' subclasses, not just constructors.

This lifts the requirement on Widgets to take an HTMLElement as its
first and only argument in the constructor.

Furthermore, the CL removes 'WidgetParams' from the generic type
parameters all-together, it was actually hiding issues as can be
seen in this CL (we were allowed to pass 'null' to a non-null setter
without complaint).

We change 'WidgetParams' to work as follows:

  * Inside the 'WidgetConfig' and 'WidgetElement' class, we use
    'Partial<WidgetT>, as it's enough for most purposes.

  * For the 'widgetConfig' function we are more strict as
    'Partial<T>' allows setting of 'undefined' or 'null' for
    properties that don't allow it.i And since we use Object.assign
    we'd set properties to 'undefined'/'null' that really shouldn't
    bet set to 'undefined'/'null'.

    We can make the "update properties" more strict by using
    'Pick<T, Keys>' and then specializing each 'widgetConfig' call
    on a concrete set of keys.
    The only oddity is that we have to intersect it with Partial<T>
    so we can pass it to the 'WidgetConfig' constructor.

R=dsv@chromium.org

Bug: None
Change-Id: If43252a646115f9a2ba68abde2f0fc0ab844a3e5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6513138
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>