[Changes] Remove sorting by file name in sidebar & combined diff view
The issue is, both CombinedDiffView and ChangesSidebar sorts the
`modifiedUISourceCodes` from `workspaceDiff` for rendering. When an
already rendered file is saved, it removes the asterisk from the title,
however it does not cause re-sorting in the changes sidebar. So, the
order of the file changes in CombinedDiffView but stays the same in
ChangesSidebar causing mismatch between where they are rendered.

This CL removes the sorting from these two places and always uses the
order provided by the `modifiedUISourceCodes`, enabling the the order in
two surfaces to be always the same.

Bug: 412968300
Change-Id: I40f63b6b9fab6e8f545d68dbd7af54de415a7511
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6519315
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>