Prepare Scripts handler for elided data source map urls
The `ScriptCatchup` v8 events have a sourceMapUrl field, which today can
contain data urls. These can be very large, especially for developer
builds. This can fill up the buffer quickly.

I'll be changing this trace event to elide sourceMapUrl if it is a data
url. To prepare for that, I created a test that will mimic the expected
trace format. Instead of the map coming from a trace event, the
Performance panel will grab it from the Script model and place it in the
metadata.

This also may be the source of some trace recordings missing data. For
example, this page[1] frequently drops trace events, perhaps because a
single trace event is too large. This page[2] has a slightly smaller
source map data url, and has no issue with recording.

[1] https://dupe-modules-lh-inline-data.surge.sh/
[2] https://dupe-modules-lh-inline-data.surge.sh/smaller

Bug: 409086863
Change-Id: If4d70d541d0e193543249aae8d3f4424868a5024
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6521671
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>