RPP: Ensure entry selection always shows details
Showing details for a time range is bound to pointerdown (as it's
listening to a drag gesture) and it's debounced from there. Showing
details for an event is bound to click.

On event selection, we could sometimes flash event details and then
replace them with the unwanted range stats.

Now we'll cancel the debounced update in those cases.

Bug:416911461
Change-Id: Ib56de828ee5286ad4f2d4b519254977e9a44ca0a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6535191
Auto-Submit: Paul Irish <paulirish@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>