Remove Recorder dependency on Timeline
If not using the revealer, relying on a direct dependency means that any
load of the Recorder panel also instantiates the Timeline (and
implicitly Network and Sources panels and perhaps more).

We roughly halve the loading time (in a debug build) by moving to the
revealer. This has a big effect in our e2e/recorder tests this panel
loading time is the largest contributor to their latency.

Change-Id: I1ba1eda82ae51bee28b138b193df10a1e8190d3a
Bug: 414578149
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6542343
Auto-Submit: Paul Irish <paulirish@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Paul Irish <paulirish@chromium.org>