[ui] Migrate EditFileSystemView to new style and add tests
This CL updates the EditFileSystemView widget from the "Settings"
panel.

The CL replaces the use of `ListWidget` with a one-column DataGrid.
Adding and editing happens in-place while deleting entries is done
via context menu.

The internal implementation of the widget also changes: No longer do
we listen for removed/added events from the model. This is not
strictly necessary since the user is the only one who can update the
excluded sub-directory list so it's enough to get the latest list of
entries when we show the widget.

Another reason is that we also have to show "invalid" entries in the
table, which are not actually synced with the model, only valid ones
are.

R=dsv@chromium.org

Fixed: 407735766
Change-Id: I4ffb7e9baf963b6e5e55ba94fc4e70e7483dc719
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6540466
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>