Fix issue with blank entries in the NetworkLogView.
When clearing the filters, it is possible (with the custom code
that deals with the waterfall column) to get into a state where
the NetworkLogView shows no rows, and it seems that all that is
needed is to trigger a redraw, because the data is available and
the grid fixes with simple acts like clicking the up and down
arrow on the scrollbar.

The tricky part with triggering an update is when to do it,
because if it is done as part of the update cycle, e.g. in
syncScrollers or redrawWaterfallColumn, it creates an infinite
update loop. By triggering it asynchronously after the filter
change has percolated, we avoid the recursion.

Fixed: 379762016
Change-Id: I07e9c42d31d0c77305f6d704569f1e5b4eb12ae3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6543876
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>