[ui] Fix 'bindToSetting' leak for boolean settings
This CL fixes the leak of the change listener in 'bindToSetting' for
boolean settings and adds tests for 'bindToSetting' as well as
'bindCheckbox'.

The CL also deprecates 'bindCheckbox' and makes it explicit that it
leaks.

For the implementation, we align 'bindCheckbox' to 'bindInput'. We
split 'bindCheckbox' into an impl function that does not use a
Setting but rather takes and returns a setter exactly mirroring
'bindInput'.

Once all usesites of 'bindCheckbox' render via lit templates and
use the 'bindToSetting' directive, we can remove 'bindCheckbox'.

R=dsv@chromium.org

Bug: 407736174
Change-Id: Iea1e756d9c2a9f34ce378dd088c249084ea28c2a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6549403
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>