[WatchBuild] Remove `--tsconfig` parameter
After crrev.com/c/6549407, main tsconfig.json is removed so watch build
was trying to use that and was failing. This CL removes the tsconfig parameter
from ESBuild and leaves it to decide which tsconfig to use.

I have tried out with e2e tests and front_end/ files and confirmed that it builds correctly.

Bug: none
Change-Id: I134329a666e11b0716366170b8215187630a7173
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6555744
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>