[MCP] Add optional selector param to styling requests
If a selector is passed to `handleMcpStylingRequest()`, the inspected
element, which is used for the context sent to the LLM, is updated to
match the selector. Without selector, the currently inspected element
is used as the context.

Element selection does not work across shadow roots or within child
frames yet.

Bug: 416134018
Change-Id: Ia53d27deb7440bdd4402676cc9fb0184c70127c3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6552400
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>