[ui] Add "Create widget" VSCode task for scaffolding widgets
After crrev.com/c/6556183 we have a way to scaffold a widget.
This CL adds a Task inside VSCode so that we can do it from
the VSCode UI.

The usage is:
* Select ">Tasks: Run Task" in the command palette.
* Select "Create Widget" task
* Supply the widget path in the prompt.
* Supply the widget name.

Then, it creates the widget for you in the given path
with the given name.

Bug: none
Change-Id: I42d6676954a3ef4d4bcf820964f9c55757013cf9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6554646
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>