[persistence] Migrate WorkspaceSettingsTab to UI vision
This CL converts `WorkspaceSettingsTab` into a presenter with a
view function. Conceptually, the view function is rather trivial:
It renders the exclude pattern regex setting and one
`EditFileSystemView` widget per file system.

Note that the number of file systems is generally rather small, so
we don't bother in maintaining a list of them in the presenter and
surgically add/remove entries. Instead, we re-calculate the full
list on every update.

Drive-by: Support RegExp settings in `bindToSetting`.

R=dsv@chromium.org

Fixed: 407736174
Change-Id: Ibfcc332f495ca58af516219183d93e39e8e138b5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6563568
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>