RPP: fix timing perf marks having AI auto-label generation
Because perf marks exist on the main thread, we were trying to generate
AI Call Trees for them even though it's not going to be very useful.
This had the effect of showing the AI auto-labels prompt on the timings
track, which we do not want.

R=andoli@chromium.org

Bug: 418223469
Change-Id: I102045bedf9fc9f3dafcb6b9c5189778d8ae6614
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6569232
Reviewed-by: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>