Convert EmptyWidget appendLink method to a setter.
This changes the EmptyWidget class to have a link property.
The reason for this is that the previous pattern of appendLink was a bit hard to use, we always have at most one link, that we need to update or remove. By making this a property, it becomes a lot simpler to manage this. It is also easier to set the link via <devtools-widget> element in a lit template.

Bug: 407751147
Change-Id: Ia2a49c3611fef7a39274cf45cda06e932da37189
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6578551
Auto-Submit: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>