[npm] Add support for fast CSS rebuilds in watch mode.
When `is_debug=true`, it is sufficient to rebuild the `.css.js` file,
which we can accomplish by asking GN for the immediate build target
and passing that to `autoninja`.

This speeds up rebuilds by a factor of 10-15x for simple CSS changes,
and basically makes the separate `npm run watch` command obsolete.

Bug: 411328609
Change-Id: Id96e447a51669def870981948fc7f071ff03f887
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6578558
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>