Fix flaky memory test
The UI rendering for the panel is throttled so it does not
happen in the same task as the input. The input while the UI
is rendering seems to be ignored, therefore, as a short-term
workaround we drain the frontend task queue to make sure
it renders in response to the input.

Fixed: 361078921
Change-Id: Idb83a0d7817fa3790fb62855bf97028a5bad20a8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6585196
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Auto-Submit: Alex Rudenko <alexrudenko@chromium.org>