[formatter] Properly reject formatter tasks that failed
Looking into a disabled web test actually discovered a bug: When we
try to substitute renamed variables and the expression fails to parse,
we silently ignore it and use the empty string as the result of the
expression substitution.

The code in the console/sources panel assumes that the formatter
throws if something goes wrong, but it never did. It just returns
`null`.

This CL updates the formatter to actually reject the Promise if the
worker runs into any trouble. All call-sites are updated to catch
Errors if there are any.

R=kimanh@chromium.org

Bug: 40267668
Change-Id: I355d9d5783a706d5c50457a32c7c1ba2c6694df4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6586566
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>