Re-enable skipped cookies test
Updated the assertion to reflect the purpose of the test
that some cookies are showing. Partitioned cookies do not
seem to be returned by CDP in this scenario. Not able
to reproduce the requestDatabaseNames error probably
due to all IndexedDb tests being disabled?

Fixed: 40911169
Change-Id: I73cbfb3aeb318d0a22477af4b0e9c4763a275c31
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6586807
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>