Migrate bottom up test to unit test
This CL migrates one of the interaction tests for the bottom up table
into a unit test.

Doing this I made the discovery that we have lots of CSS in
TimelinePanel for the data grid, so I moved that into a new file. This
fixes how the table looks in the screenshot.

I also created a helper to render a widget within a VBox, as we do that
a lot in RPP in the summary drawer and it's required to have the child
fill the space correctly.

Finally, I also deleted the redundant "long task candystripe" test, as
the FlameChart.test.ts file already contains decoration unit tests, so
we do not need to retest it.

Bug: 413413595
Change-Id: I1363f3cff1a4f5b3e93d3ac2e92937572b8c6597
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6582806
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>