Migrate call tree test from interaction => unit
This CL adds a screenshot to an existing test for the construction of
the call tree, and removes the old interaction test. It also removes an
interaction test that screenshots the entire panel; we do not need this
test as we want to test via e2e tests for the overall panel (which
already exist) and unit tests for smaller parts of the panel (which we
are working on as part of this migration).

Follow-up CLs will migrate more of timeline_test.ts into e2e and/or unit
tests.

R=finnur@chromium.org

Bug: 413413595
Change-Id: I62670a40533e3c7ae42cc350b0fb3fc424eaa895
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6589777
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>