Fix window size in unit tests
The window size was not fixed for unit tests
resulting in a different viewport being used when tests
run in headless. Since our tests run in an iframe, rendering
something that exceeds the iframe height would cause the bounding
box of the elements becoming larger that the viewport resulting
in screenshots containing parts of the parent iframe.

Bug: none
Change-Id: Ifee7587e67a5e68aedee572cfaa0fa18a4a54bd5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6597734
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>