Look for stack frames to parse without looking for Error
Previously we searched for a message to start with a word
ending in "Error" to determine if it was an error and we
should look for stack frames. With this change we make no
assumptions about what an error is named and just look for
stack frames instead.

Bug: 406029201
Change-Id: I7c31c6b4cd4a9137227a99baf34626cdd51dbc16
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6563869
Commit-Queue: Eric Leese <leese@chromium.org>
Auto-Submit: Eric Leese <leese@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>