Fix a flaky coverage test on Linux
The issue was due to the height of the coverage panel on Linux being
not enough to show all the rows, and due to the inner rendering
mechanism of data grids, hidden rows are actually also hidden from the
DOM tree. This CL ensures that the coverage panel is at the top, which
gives it enough height to show all the rows.

Fixed: 40264187
Change-Id: I72e10476c4944170d1891f094a23d2c0c74fc0af
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6604062
Auto-Submit: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>