tests: avoid `document.body` and update lint rule
This CL updates the existing ESLint rule that checks for
`widget.show(document.body)` and matches `document.body.append` and
`appendChild`.

Note: the rule should be renamed, but this CL is large enough already so
I will do that in a follow-up.

Making this change also highlighted a bug where we relied on the
`document.body.removeChildren` call in `ThemeSupport.test.ts` to tidy up
DOM; I tracked this down to the CodeMirror tooltip and added code in
those tests to remove it (not sure why CM doesn't auto destroy it).

Bug: 421811012
Change-Id: I7635e72504ac4f98ff128039f37dbf8427bb0cbc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6611218
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>