[e2e] Port console filter resetting test to non hosted mode
Note that since the test is flaky we don't port it 1:1. The
'getConsoleMessages' helper is rather terrible: It also navigates, it
has a random 100ms timeout and then just reads the console messages
that are there at that point in time.

Instead, the new test waits until the expected number of messages
have appeared. Filters, waits for the new number of messages, removes
the filter and check that we are back to the original number.

This should be more robust, and more importantly: no more random
timeout.

Bug: 416522177
Change-Id: I94e5f0e6657328d568725960f8766a4a68e511cd
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6615239
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>