Enable the NetworkDependencyTree unittest
The test looks good, and the only potential timeout reason is loading the trace,
But the trace loaded for this test is pretty small, so it should be fine.

Also checked the history run, it seems only failed twice.

So tentative to re-enable the test to see if it fails again.

Bug: 403507404
Change-Id: I447df9562e81bdc8be4459bc5111714025d56b3b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6619071
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>