[Lantern] Prevent graph cycles from redirects
`createGraph` sets the requestedUrl based on the first Lantern request.
DevTools requests have to be mapped to Lantern requests. For redirects,
DevTools keeps everything in one request object, but Lantern wants them
as separate request objects.

This process turned a request with redirects from this:

- devtools request (final)
  - redirect source 1
  - redirect source 2
  - redirect source 3

(representing source 1 -> 2 -> 3 -> final)

into this for Lantern:

- lantern request (final)   <- this should be last!
- redirect source 1         <- the actual initial nav url
- redirect source 2
- redirect source 3

This resulted in an incorrect selection for the requestedUrl.
`createGraph` grabbed the final redirected URL, instead of the initial
requested URL. Lantern uses that as the root request, which for this
above case resulted in a cycle in the page graph.

Fixed: 405979476
Change-Id: Ic2db2a61e028bf0069724fc284a030af6f2bd226
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6626095
Auto-Submit: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>