[e2e] Use command palette to open Lighthouse panel
The previous code was trying to click on the tab itself, but the
fallback code for when the tab is hidden didn't seem to ever work. So
when the width of the test runner window is small enough, Lighthouse
tests would always timeout.

Bug: 422939900
Change-Id: If006614242534de39702c1e1d70eea7bdaa1b631
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6628535
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>