RPP: use "local metrics" and "field metrics" consistently
This changes it because before we had "Live metrics" and "Field data" as
the prominent terms, and that's a bit of a mixup.

This CL makes us consistently use:

- Local metrics
- Field metrics

Which is consistent; and avoids any confusion of "live metrics" which is
not as clear as "Local metrics".

To avoid confusion regarding the title of the view itself, we keep the
title of the page to "Live metrics". So the way to think about this is:

1. We have a **Live metrics** page
2. That shows both **local metrics** and **field metrics** (if
   configured).

(Thanks to @barrypollard for his input).

R=paulirish@chromium.org

Fixed: 404513461
Change-Id: Ief19e37ac4968d4fe47a6e58c40983dfeefd8986
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6624111
Reviewed-by: Paul Irish <paulirish@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>