Reland - Make drawer orientation configurable
Made sure the feature is hided behind an experiment for the reland

Explainer:
https://docs.google.com/document/d/1oVBzqCC-t9nVVs5crNI8QHP2MFucTg5E7LSMjKAnajM/

This CL makes drawer orientation configurable. The user will be able to
toggle the orientation of the drawer between horizontal or vertical via
the button next to closing drawer or pressing "shift + esc".

Bug: 416086584
Change-Id: I6c0313299efaff55a52f6a8a1ddc6d7e17e874b0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6548645
Commit-Queue: Jayson Chen <jaysonchen@microsoft.com>
Reviewed-by: Robert Paveza <Rob.Paveza@microsoft.com>
Reviewed-by: Simon Zünd <szuend@chromium.org>