Remove usages in test of `addSourceMapping`
The `addSourceMapping` was previously added for source formatting.
While it is no longer needed, tests have used it as a workaround
to map scripts to ui locations. Update the tests to make use of the
`MockProtocolBackend` instead.

In a follow-up, we'll remove the code around `addSourceMapping`.

Bug: 40262699
Change-Id: I1123084a298d5a354f6f563122237ac6c0a78558
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6633971
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>