[Changes] Inline `patching` flag as `true` in Changes panel
Drive-by:
* Remove `Toolbar` when the `CombinedDiffView` is
rendered. It renders a `revert` button that does nothing,
so we can remove the toolbar altogether.
(Checked this with Peter too)
* Removed a TODO that would implement
clicking on a line and jumping to it in the
sources panel. Considering that it's a nice to have
feature that is not planned to be implemented
for CombinedDiffView.

Fixed: 412968300
Change-Id: Ifba2df42406ca830e043d6aa687be7fc70e8332a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6632913
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>