Center outline of icon on focus
This CL moves the outline for icon buttons to the icon itself, in an
attempt to properly center the focus outline with the icon. Since
we have different sizes for icons and buttons, we need to adjust the
outline for each of the size.

Fixed: 390136819
Change-Id: Ibaa53c5419e458d0ef8f3e82c13f051935dde657
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6633616
Commit-Queue: Kateryna Prokopenko <kprokopenko@chromium.org>
Reviewed-by: Kateryna Prokopenko <kprokopenko@chromium.org>
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>