RPP: ensure we handle unexpected non-string props in e11y events
This fixes a bug where a `null` value was set for a property and failed
to render the summary view, breaking the ability to click + view
information about a perf.mark.

This CL fixes the call site by handling non-string data and updating the
types to set the value to `unknown`. We only support strings in the API
officially, but we should make sure that unexpected data does not cause
a runtime error.

R=finnur@chromium.org

Fixed: 424369233
Change-Id: I6f329600b2d76c3f47d05d2e7aab65c7e821acf1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6633605
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>