[RPP] Disable insights tab if there are no insight sets
I don't have a repro for this. I noticed via inspection of the code that
it may be possible for an empty insights map to be generated. To defend
against that, now using the size of the insights map to determine
whether to disable the insights tab.

Bug: None
Change-Id: I38c253d1a7095fab3a631cc7ae179b8791033938
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6643512
Auto-Submit: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>