Fix keyboard shortcuts in layers panel
Keyboard shortcuts weren't working, as the drag handler within the
TransformController (for drag and drop) prevented the default action
to happen on mouse down, including moving the focus to the element that
was dragged. As a result, the focus never got to the layers 3d view and
instead was wrongly set to other parts of the layers panel. Without
having focus, the keyboard listeners didn't receive any keyboard
events.

This CL optionally allows default actions to happen on mouse down.

Fixed: 366067954
Change-Id: I3c15f300f282043cb3fd5840c241f230430cf276
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6651066
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>